+++ docs/scripting/examples/custom-actions.qml
@@ -21,6 +21,8 @@ QtObject {
 
         // create a menu entry to transform text with rot13
         script.registerCustomAction("transformTextRot13", "Transform selected text with rot13", "rot13", "text-wrap", true);
+
+        script.registerCustomAction("noteSubFolder", "Show active note subfolder information", "Subfolder");
     }
 
     /**
@@ -73,6 +75,13 @@ QtObject {
                 });
 
                 script.noteTextEditWrite(text);
+                break;
+
+            case "noteSubFolder":
+                var noteSubFolderQmlObj = Qt.createQmlObject("import QOwnNotesTypes 1.0; NoteSubFolder{}", mainWindow, "noteSubFolder");
+                var subFolder = noteSubFolderQmlObj.activeNoteSubFolder();
+                script.log(subFolder.fullPath());
+                script.log(subFolder.relativePath());
                 break;
         }
     }